manage_telemetry_db/requirements.txt
requests
pyproj
dotenv
configparser
psycopg2-binary